rename: 패키지명 수정
- table to tablegroup
