+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -5,8 +5,8 @@ import java.util.Objects;
 import kitchenpos.domain.order.OrderValidator;
 import kitchenpos.domain.ordertable.OrderTable;
 import kitchenpos.domain.ordertable.OrderTableRepository;
-import kitchenpos.domain.table.TableGroup;
-import kitchenpos.domain.table.repository.TableGroupRepository;
+import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.domain.tablegroup.repository.TableGroupRepository;
 import kitchenpos.dto.request.TableGroupCreateRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
