+++ src/main/java/kitchenpos/domain/tablegroup/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.domain.table.repository;
+package kitchenpos.domain.tablegroup.repository;
 
-import kitchenpos.domain.table.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
