+++ src/main/java/kitchenpos/dto/response/TableGroupResponse.java
@@ -2,7 +2,7 @@ package kitchenpos.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.domain.table.TableGroup;
+import kitchenpos.domain.tablegroup.TableGroup;
 
 public class TableGroupResponse {
